pyproject.toml
[project]
name = "thesis"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "asttokens==3.0.0",
    "attrs==25.3.0",
    "backoff==2.2.1",
    "black==25.1.0",
    "certifi==2025.1.31",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "comm==0.2.2",
    "contourpy==1.3.1",
    "cycler==0.12.1",
    "debugpy==1.8.12",
    "decorator==5.2.1",
    "deeptime==0.4.5",
    "distro==1.9.0",
    "einops==0.8.1",
    "eval-type-backport==0.2.2",
    "executing==2.2.0",
    "fastjsonschema==2.21.1",
    "filelock==3.19.1",
    "fonttools==4.56.0",
    "fsspec==2025.2.0",
    "gitdb==4.0.12",
    "gitpython==3.1.45",
    "h11==0.16.0",
    "hf-xet==1.1.10",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "huggingface-hub==0.35.3",
    "idna==3.10",
    "iniconfig==2.3.0",
    "ipykernel==6.29.5",
    "ipython==9.0.1",
    "ipython-pygments-lexers==1.1.1",
    "ipywidgets==8.1.5",
    "jedi==0.19.2",
    "jinja2==3.1.5",
    "jiter==0.10.0",
    "joblib==1.4.2",
    "jsonschema==4.25.0",
    "jsonschema-specifications==2025.4.1",
    "jupyter-client==8.6.3",
    "jupyter-core==5.7.2",
    "jupyterlab-widgets==3.0.13",
    "kiwisolver==1.4.8",
    "llvmlite==0.44.0",
    "markupsafe==3.0.2",
    "matplotlib==3.10.0",
    "matplotlib-inline==0.1.7",
    "mpmath==1.3.0",
    "mypy-extensions==1.0.0",
    "narwhals==2.1.1",
    "nbformat==5.10.4",
    "nest-asyncio==1.6.0",
    "networkx==3.4.2",
    "numba==0.61.0",
    "numpy==2.1.3",
    "nvidia-cublas-cu12==12.4.5.8",
    "nvidia-cuda-cupti-cu12==12.4.127",
    "nvidia-cuda-nvrtc-cu12==12.4.127",
    "nvidia-cuda-runtime-cu12==12.4.127",
    "nvidia-cudnn-cu12==9.1.0.70",
    "nvidia-cufft-cu12==11.2.1.3",
    "nvidia-curand-cu12==10.3.5.147",
    "nvidia-cusolver-cu12==11.6.1.9",
    "nvidia-cusparse-cu12==12.3.1.170",
    "nvidia-cusparselt-cu12==0.6.2",
    "nvidia-nccl-cu12==2.21.5",
    "nvidia-nvjitlink-cu12==12.4.127",
    "nvidia-nvtx-cu12==12.4.127",
    "openai==1.88.0",
    "packaging==24.2",
    "pandas==2.2.3",
    "parso==0.8.4",
    "pathspec==0.12.1",
    "pexpect==4.9.0",
    "pillow==11.1.0",
    "platformdirs==4.4.0",
    "plotly==6.3.0",
    "pluggy==1.6.0",
    "posthog==6.7.6",
    "prompt-toolkit==3.0.50",
    "protobuf==6.33.1",
    "psutil==7.0.0",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.11.0",
    "pygments==2.19.1",
    "pymde==0.1.18",
    "pynndescent==0.5.13",
    "pyparsing==3.2.1",
    "pyperclip==1.11.0",
    "pytest==9.0.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "pytz==2025.2",
    "pyyaml==6.0.3",
    "pyzmq==26.2.1",
    "referencing==0.36.2",
    "requests==2.32.5",
    "rpds-py==0.27.0",
    "scikit-learn==1.6.1",
    "scipy==1.13.0",
    "seaborn==0.13.2",
    "sentry-sdk==2.45.0",
    "setuptools==75.8.2",
    "six==1.17.0",
    "smmap==5.0.2",
    "sniffio==1.3.1",
    "stack-data==0.6.3",
    "sympy==1.13.1",
    "tabpfn==2.2.1",
    "tabpfn-common-utils==0.2.2",
    "tabulate==0.9.0",
    "threadpoolctl==3.5.0",
    "torch==2.6.0",
    "torchvision==0.21.0",
    "tornado==6.4.2",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "triton==3.2.0",
    "typing-extensions==4.12.2",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "urllib3==2.3.0",
    "wandb==0.23.0",
    "wcwidth==0.2.13",
    "widgetsnbextension==4.0.13",
]
